Description: Sim needs timetable and shunting
Sim requires timetable.csv and shunting.csv to function. These need to be files of length 672 (1 week in 15 minute slows).
Timetable represents the requirement for trains from the depot.
Injection is either TRUE or FALSE. When TRUE, the maintenance activity can NOT start - as trains are currently being "injected" into service.